services/notification-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
httpx==0.27.2
bcrypt==4.2.0
PyJWT==2.9.0
prometheus-client==0.21.0
python-multipart==0.0.12
psycopg2-binary==2.9.9
